pyproject.toml
[project]
name = "bejo-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.4",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.4.7",
    "loguru>=0.7.3",
    "mem0ai>=0.1.101",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
